Refactored logic concerning remote charge points
